refactor: separate app initialization, add router
